feat(runtime): close the remaining gaps against the merged iOS loader PR
Four features from the merged overhaul had not made it into the port,
plus the structural move that belonged with them:

- ES modules use the compiled-code cache: CompileFileEsModule consumes
  and produces cache blobs through the existing scheme under the same
  config gate, with module caches keyed as .mcache - a classic require()
  and an import of the same .js file each keep their own blob instead of
  overwriting each other's on every load.
- Concurrent module fetches are capped at 16 process-wide: excess jobs
  queue and finishing threads drain them, so the cap bounds threads (and
  JVM attaches), not just sockets, and the caller never blocks. A failed
  thread spawn delivers a transport-error completion instead of wedging
  the graph pump, and jobs queued behind it fail rather than wait for a
  thread that will never exist.
- The transport takes canonical keys from its callers - computed on the
  isolate thread, off-thread canonicalization impossible by construction.
  This also fixes cache-bust marks for collapsed-scheme URLs: eviction
  marks under the repaired registry key, and the transport previously
  canonicalized the raw URL, so the mark could never match or clear.
- Workers accept http(s) entries: the constructor classifies the URL up
  front and skips filesystem resolution, the existing HTTP entry branch
  and boot options do the work, and the settle gate probes the same
  canonical key the entry registers under. The worker inspector target
  passes an http(s) URL through instead of prefixing file://.
- The ns:module binding lives beside the loader state it configures; the
  transport TU no longer depends on the loader headers, and
  configureLoader parses the import map once and installs the parsed
  result instead of validating by parse and parsing again.
